+++ option.go
@@ -41,3 +41,21 @@ type (
 	Option  func(*options)
 	options Config
 )
+
+// WithValueFormatter provides the value formatter for Config.Explain.
+// It's for hiding sensitive information (e.g. password, secret) which should not be exposed.
+//
+// By default, it uses fmt.Sprint to format the value.
+func WithValueFormatter(valueFormatter func(path string, loader Loader, value any) string) ExplainOption {
+	return func(options *explainOptions) {
+		options.valueFormatter = valueFormatter
+	}
+}
+
+type (
+	// ExplainOption configures Config.Explain with specific options.
+	ExplainOption  func(*explainOptions)
+	explainOptions struct {
+		valueFormatter func(path string, loader Loader, value any) string
+	}
+)